Inlogsysteem, en afscherming  afgewerkt. Uitlening: uitlenen, terugbrengen, verlengen afgewerkt

+++ DeKrekelGroup5/DeKrekelGroup5/Models/Domain/Instellingen.cs
@@ -10,7 +10,7 @@ namespace DeKrekelGroup5.Models.Domain
     {
         [Key]
         public int Id { get; set; }
-        public decimal BedragBoetePerDag { get; set; }
+        public int BedragBoetePerDag { get; set; }
         public int UitleenDagen { get; set; }
         public int MaxVerlengingen { get; set; }
 
@@ -19,7 +19,7 @@ namespace DeKrekelGroup5.Models.Domain
             
         }
 
-        public Instellingen(int id, decimal bedragBoetePerDag, int uitleenDagen, int maxVerlengingen)
+        public Instellingen(int id, int bedragBoetePerDag, int uitleenDagen, int maxVerlengingen)
         {
             Id = id;
             BedragBoetePerDag = bedragBoetePerDag;
